services/auth/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
databases[postgresql]
asyncpg
psycopg2
